+++ src/main/java/bl4ckscor3/mod/theplopper/Configuration.java
@@ -2,15 +2,10 @@ package bl4ckscor3.mod.theplopper;
 
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.Config.Comment;
-import net.minecraftforge.common.config.Config.RangeInt;
 
 @Config(modid=ThePlopper.MOD_ID, category="general")
 public class Configuration
 {
-	@Comment("The range in blocks at which the plopper will still pick up items")
-	@RangeInt(min=1, max=20)
-	public static byte range = 3;
-
 	@Comment("Whether particles should show up every time the plopper picks up an item")
 	public static boolean displayParticles = true;