+++ src/main/java/bl4ckscor3/mod/theplopper/gui/GuiPlopper.java
@@ -15,6 +15,9 @@ public class GuiPlopper extends GuiContainer
 	public GuiPlopper(InventoryPlayer playerInv, TileEntityPlopper te)
 	{
 		super(new ContainerPlopper(playerInv, te));
+
+		xSize = 200;
+		ySize = 133;
 	}
 
 	@Override
@@ -24,7 +27,7 @@ public class GuiPlopper extends GuiContainer
 
 		String name = I18n.format("tile.theplopper:plopper.name");
 
-		fontRenderer.drawString(name, xSize / 2 - fontRenderer.getStringWidth(name) / 2, 7, 0x404040);
+		fontRenderer.drawString(name, (xSize - 24) / 2 - fontRenderer.getStringWidth(name) / 2, 7, 0x404040);
 	}
 
 	@Override