+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-# v1.1
+# v1.2
+- Add range upgrade
+- Removed range config option
 
+# v1.1.1
+- Prevent crash with VanillaFix
+
+# v1.1
 - Fix items stacking in the plopper's inventory when they had a different metadata but the same item ID
 
-
 # v1.0
-
 - Initial release